Refactor codes of resource queue relating to memory policy. This is to make it easier to add in memory policy codes for resource group.
