+++ src/backend/utils/resource_manager/Makefile
@@ -12,6 +12,6 @@ subdir = src/backend/utils/resource_manager
 top_builddir = ../../../..
 include $(top_builddir)/src/Makefile.global
 
-OBJS = resource_manager.o
+OBJS = resource_manager.o memquota.o
 
 include $(top_srcdir)/src/backend/common.mk
