+++ src/backend/utils/resscheduler/Makefile
@@ -12,6 +12,6 @@ subdir = src/backend/utils/resscheduler
 top_builddir = ../../../..
 include $(top_builddir)/src/Makefile.global
 
-OBJS = resqueue.o resscheduler.o memquota.o
+OBJS = resqueue.o resscheduler.o
 
 include $(top_srcdir)/src/backend/common.mk
