+++ src/include/utils/resscheduler.h
@@ -172,4 +172,18 @@ extern void AtAbort_ResScheduler(void);
 extern void ResHandleUtilityStmt(Portal portal, Node *stmt);
 extern bool ResLockUtilityPortal(Portal portal, float4 ignoreCostLimit);
 
+ /**
+  * What is the memory limit on a queue per the catalog in bytes. Returns -1 if not set.
+  */
+extern int64 ResourceQueueGetMemoryLimitInCatalog(Oid queueId);
+
+/**
+ * What is the memory limit configuration for a specific queue?
+ */
+extern int64 ResourceQueueGetMemoryLimit(Oid queueId);
+
+/*
+ * How many memory reserved for the query
+ */
+extern uint64 ResourceQueueGetQueryMemoryLimit(PlannedStmt *stmt, Oid queueId);
 #endif   /* RESSCHEDULER_H */
